pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comiccatcher"
dynamic = ["version"]
authors = [
  { name = "Tony Beville" },
]
description = "A high-performance native desktop OPDS 2.0 comic reader"
readme = "DESIGN.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyQt6>=6.7.0",
    "qasync>=0.27.1",
    "httpx>=0.27.0",
    "pydantic>=2.6.4",
    "comicbox>=2.2.1",
    "Pillow>=10.2.0",
    "rarfile>=4.2",
    "py7zr>=1.1.0",
    "PyMuPDF>=1.24.0",
]

[project.scripts]
comiccatcher = "comiccatcher.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
comiccatcher = ["resources/*", "resources/icons/*"]

[tool.setuptools.dynamic]
version = {attr = "comiccatcher.__version__"}
